BPE26-ref: extract the track title operations to dedicated file
